refactor(types): reorganize Firebase type declarations and enhance TypeScript support
- Updated `tsconfig.json` to include `.d.ts` files for improved type recognition.
- Moved Firebase authentication type declarations to a new `firebase-mocks.d.ts` file for better organization and maintainability.
- Removed redundant type declarations from `declarations.d.ts`, streamlining the codebase.
